Adding mute commands and giveawayban

+++ events/giveawayReactRemove.js
@@ -8,6 +8,12 @@ module.exports = {
 		if (user.bot) return;
 		const giveaway = db.getData("/giveaways").find((giveaway) => giveaway.msg_id == reaction.message.id);
 		if (!giveaway) return;
+
+		// Cheks if the member was banned from giveaways
+		const banned_users = db.getData("/giveawaysban");
+		if (banned_users.length > 0 && banned_users.includes(user.id)) return;
+		// Save user leaving
+
 		log(`User ${user.username} leave giveaway ${giveaway.id} !`);
 		db.delete(
 			`/giveaways[${db.getIndex("/giveaways", giveaway.id, "id")}]/participants[${db.getIndex(
